--- celesta-test/src/main/java/org/testcontainers/containers/MSSQLServerContainerProvider.java
@@ -1,13 +1,0 @@
-package org.testcontainers.containers;
-
-public class MSSQLServerContainerProvider extends JdbcDatabaseContainerProvider {
-    @Override
-    public boolean supports(String databaseType) {
-        return databaseType.equals(MSSQLServerContainer.NAME);
-    }
-
-    @Override
-    public JdbcDatabaseContainer newInstance(String tag) {
-        return new MSSQLServerContainer(MSSQLServerContainer.IMAGE + ":" + tag);
-    }
-}
