--- celesta-test/src/test/resources/META-INF/services/org.testcontainers.containers.JdbcDatabaseContainerProvider
@@ -1,1 +1,0 @@
-org.testcontainers.containers.MSSQLServerContainerProvider
