+++ celesta-test/src/test/resources/testcontainers.properties
@@ -1,0 +1,1 @@
+oracle.container.image=thebookpeople/oracle-xe-11g
